requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
psycopg2-binary==2.9.7
pandas==2.1.4
schedule==1.2.0
python-dotenv==1.0.0
lxml==4.9.3
html5lib==1.1
urllib3==2.0.7
